requirements.txt
django-heroku==0.3.1
gunicorn==19.9.0
whitenoise==4.1.3
social-auth-app-django==3.1.0
python-social-auth==0.3.6